+++ peering/tests/test_functions.py
@@ -2,12 +2,12 @@ from unittest.mock import patch
 
 from django.test import TestCase
 
-from .. import call_irr_as_set_resolver, parse_irr_as_set
+from ..functions import call_irr_as_set_resolver, parse_irr_as_set
 from .mocked_data import *
 
 
 class IRRASSetFunctions(TestCase):
-    @patch("peering.subprocess.Popen", side_effect=mocked_subprocess_popen)
+    @patch("peering.functions.subprocess.Popen", side_effect=mocked_subprocess_popen)
     def test_call_irr_as_set_resolver(self, mocked_popen):
         prefixes6 = call_irr_as_set_resolver("AS-MOCKED")
         self.assertEqual(1, len(prefixes6))
